CXX-1045 Replace hint::to_document() with hint::to_value()
The design of hint::to_document() is flawed in that it returns a
document with a fixed key name of "$hint", which doesn't always
represent the hint option correctly for database operations (commands
take this option with a key name of "hint", for example).

To address this issue, this commit removes knowledge of the key name
from the hint class, by replacing hint::to_document() with
hint::to_value().